Overhauled CSS
Improved the looks and am almost done with the view. There are more input elements that will be added later but the view is able to resize itself appropriatly. The only thing left to make responsive is the harmonic table itself - I haven't quite figured out how to resize those hexagons based on the height of its parent div